Backend/requirements/test.txt
-r base.txt

pytest==2.7.2
pytest-django==2.8.0
factory-boy==2.5.2
